frontend/backend/requirements.txt
asgiref==3.12.0
Django==6.1.1
django-cors-headers==4.10.0
djangorestframework==3.18.0
sqlparse==0.5.5
tzdata==2026.2
